# src/models/API.ts
/**
 * This file was auto-generated by swagger-to-ts.
 * Do not make direct changes to the file.
 */

export namespace API {
  export interface TrackerRank {
    /**
     * Tracker ID
     */
    tracker_uid: number;
    /**
     * Rank for maximum speed of tracker
     */
    max_speed_rank: number;
  }
  export interface Event {
    uid: number;
    tracker_uid: number;
    angle: number;
    speed: number;
    /**
     * Timestamp
     */
    aquisition_time: number;
    visible_satellites: number;
    engine: 'on' | 'off';
    event_id: number;
    event_info: number;
    insert_time: string;
    mileage: number;
    voltage: number;
    driver_ibutton: string;
    hdop: number;
  }
  export interface ErrorResponse {
    /**
     * Error message
     */
    message: string;
    code: 'VALIDATION_ERROR' | 'INTERNAL_ERROR';
  }
}